Update .gitignore with Visual Studio and build files
Updated the `.gitignore` file to include several new lines. These changes include a comment block indicating that the file was automatically created by Microsoft Visual Studio, and lines to ignore the `.vs` directory and the `obj` directory within the `HackathonApp` directory. These directories typically contain user-specific and temporary build files that are not necessary to include in the repository.